third_party/blink/web_tests/external/wpt/css/css-values/calc-infinity-nan-serialize-length-expected.txt

This is a testharness.js-based test.
PASS 'calc(1px * NaN)' as a specified value should serialize as 'calc(NaN * 1px)'.
FAIL 'calc(1in * NaN)' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1in * NaN)' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(NaN * 1in)"
FAIL 'calc(1cm * NaN)' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1cm * NaN)' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(NaN * 1cm)"
FAIL 'calc(1mm * NaN)' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1mm * NaN)' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(NaN * 1mm)"
FAIL 'calc(1q * NaN)' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1q * NaN)' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(NaN * 1q)"
FAIL 'calc(1pt * NaN)' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1pt * NaN)' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(NaN * 1pt)"
FAIL 'calc(1pc * NaN)' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1pc * NaN)' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(NaN * 1pc)"
PASS 'calc(1px * nan)' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * infinity / infinity)' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * 0 * infinity)' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * (infinity + -infinity))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * (-infinity + infinity))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * (infinity - infinity))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * infinity)' as a specified value should serialize as 'calc(infinity * 1px)'.
PASS 'calc(1px * -infinity)' as a specified value should serialize as 'calc(-infinity * 1px)'.
PASS 'calc(1px * iNFinIty)' as a specified value should serialize as 'calc(infinity * 1px)'.
PASS 'calc(1px * (infinity + infinity))' as a specified value should serialize as 'calc(infinity * 1px)'.
PASS 'calc(1px * (-infinity + -infinity))' as a specified value should serialize as 'calc(-infinity * 1px)'.
PASS 'calc(1px * 1/infinity)' as a specified value should serialize as 'calc(0px)'.
PASS 'calc(1px * infinity * infinity)' as a specified value should serialize as 'calc(infinity * 1px)'.
PASS 'calc(1px * -infinity * -infinity)' as a specified value should serialize as 'calc(infinity * 1px)'.
FAIL 'calc(1 * max(INFinity*3px, 0px))' as a specified value should serialize as 'calc(infinity * 1px)'. assert_equals: 'calc(1 * max(INFinity*3px, 0px))' and 'calc(infinity * 1px)' should serialize the same in specified values. expected "calc(infinity * 1px)" but got "calc(1 * max(infinity * 1px, 0px))"
FAIL 'calc(1 * min(inFInity*4px, 0px))' as a specified value should serialize as 'calc(0px)'. assert_equals: 'calc(1 * min(inFInity*4px, 0px))' and 'calc(0px)' should serialize the same in specified values. expected "calc(0px)" but got "calc(1 * min(infinity * 1px, 0px))"
FAIL 'calc(1 * max(nAn*2px, 0px))' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1 * max(nAn*2px, 0px))' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(1 * max(NaN * 1px, 0px))"
FAIL 'calc(1 * min(nan*3px, 0px))' as a specified value should serialize as 'calc(NaN * 1px)'. assert_equals: 'calc(1 * min(nan*3px, 0px))' and 'calc(NaN * 1px)' should serialize the same in specified values. expected "calc(NaN * 1px)" but got "calc(1 * min(NaN * 1px, 0px))"
FAIL 'calc(1 * clamp(-INFINITY*20px, 0px, infiniTY*10px))' as a specified value should serialize as 'calc(0px)'. assert_equals: 'calc(1 * clamp(-INFINITY*20px, 0px, infiniTY*10px))' and 'calc(0px)' should serialize the same in specified values. expected "calc(0px)" but got "calc(1 * clamp(-infinity * 1px, 0px, infinity * 1px))"
PASS 'calc(1px * max(NaN, min(0,10)))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * clamp(NaN, 0, 10))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * max(0, min(10, NaN)))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * clamp(0, 10, NaN))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * max(0, min(NaN, 10)))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * clamp(0, NaN, 10))' as a specified value should serialize as 'calc(NaN * 1px)'.
PASS 'calc(1px * clamp(-Infinity, 0, infinity))' as a specified value should serialize as 'calc(0px)'.
PASS 'calc(1px * clamp(-inFinity, infinity, 10))' as a specified value should serialize as 'calc(10px)'.
Harness: the test ran to completion.
